# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto Android Laravel",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
        "/mesasPorPiso/{tiempoToken}/{piso}": {
            "get": {
                "description": "Mesas por cada piso",
                "operationId": "mesasPorPiso",
                "parameters": [
                    {
                        "name": "tiempoToken",
                        "in": "path",
                        "description": "Tiempo del token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "piso",
                        "in": "path",
                        "description": "Numero de piso",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Id del pedido."
                    }
                }
            }
        },
        "/crearPedido": {
            "post": {
                "description": "Crear pedido",
                "operationId": "crearPedido",
                "parameters": [
                    {
                        "name": "idMesa",
                        "in": "path",
                        "description": "Id mesa",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    },
                    {
                        "name": "idCreador",
                        "in": "path",
                        "description": "Id creador pedido",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Id del pedido."
                    }
                }
            }
        },
        "/PedidoDetallePlatos/{tiempoToken}/{idPedido}": {
            "get": {
                "description": "Detalles del pedido",
                "operationId": "PedidoDetallePlatos",
                "parameters": [
                    {
                        "name": "tiempoToken",
                        "in": "path",
                        "description": "Tiempo del token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "idPedido",
                        "in": "path",
                        "description": "Id pedido",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Detalles pedido."
                    }
                }
            }
        },
        "/eliminarPlatoDetalle": {
            "delete": {
                "description": "Eliminar detalles del plato",
                "operationId": "eliminarPlatoDetalle",
                "parameters": [
                    {
                        "name": "idDetalle",
                        "in": "path",
                        "description": "Id del detalles del plato",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Se ha eliminado."
                    }
                }
            }
        },
        "/crearDetalle": {
            "get": {
                "description": "Crear detalles del pedido",
                "operationId": "crearDetalle",
                "parameters": [
                    {
                        "name": "idPedido",
                        "in": "path",
                        "description": "Id pedido",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    },
                    {
                        "name": "idPlato",
                        "in": "path",
                        "description": "Id plato",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    },
                    {
                        "name": "precio",
                        "in": "path",
                        "description": "precio del plato",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    },
                    {
                        "name": "idPedido",
                        "in": "path",
                        "description": "Id del mesero el cual realza el pedido",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Se ha agregado."
                    }
                }
            }
        },
        "/listaPisos/{tiempoToken}": {
            "get": {
                "description": "Listar los pisos",
                "operationId": "listaPisos",
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Datos pisos."
                    }
                }
            }
        },
...
    }
}